Remove deprecated collector
