+++ internal/fullnode/peer_collector_test.go
@@ -5,115 +5,197 @@ import (
 	"errors"
 	"testing"
 
+	"github.com/cometbft/cometbft/p2p"
 	cosmosv1 "github.com/strangelove-ventures/cosmos-operator/api/v1"
-	"github.com/strangelove-ventures/cosmos-operator/internal/cosmos"
 	"github.com/stretchr/testify/require"
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-type mockLister func(ctx context.Context, list client.ObjectList, opts ...client.ListOption) error
+type mockGetter func(ctx context.Context, key client.ObjectKey, obj client.Object, opts ...client.GetOption) error
 
-func (fn mockLister) List(ctx context.Context, list client.ObjectList, opts ...client.ListOption) error {
+func (fn mockGetter) Get(ctx context.Context, key client.ObjectKey, obj client.Object, opts ...client.GetOption) error {
 	if ctx == nil {
-		panic("context is nil")
+		panic("nil context")
 	}
-	return fn(ctx, list, opts...)
+	if len(opts) > 0 {
+		panic("unexpected opts")
+	}
+	return fn(ctx, key, obj, opts...)
 }
 
-type mockTmClient func(ctx context.Context, rpcHost string) (cosmos.TendermintStatus, error)
+var panicGetter = mockGetter(func(ctx context.Context, key client.ObjectKey, obj client.Object, opts ...client.GetOption) error {
+	panic("should not be called")
+})
 
-func (fn mockTmClient) Status(ctx context.Context, rpcHost string) (cosmos.TendermintStatus, error) {
-	return fn(ctx, rpcHost)
-}
-
-func TestPeerCollector_CollectAddresses(t *testing.T) {
+func TestPeerCollector_Collect(t *testing.T) {
 	t.Parallel()
 
 	ctx := context.Background()
+	const (
+		namespace = "strangelove"
+		nodeKey   = `{"priv_key":{"type":"tendermint/PrivKeyEd25519","value":"HBX8VFQ4OdWfOwIOR7jj0af8mVHik5iGW9o1xnn4vRltk1HmwQS2LLGrMPVS2LIUO9BUqmZ1Pjt+qM8x0ibHxQ=="}}`
+	)
 
-	pods := []corev1.Pod{
-		{ObjectMeta: metav1.ObjectMeta{Name: "pod-0"}, Status: corev1.PodStatus{PodIP: "1.1.1.1"}},
-		{ObjectMeta: metav1.ObjectMeta{Name: "pod-1"}, Status: corev1.PodStatus{PodIP: "2.2.2.2"}},
-	}
-	var crd cosmosv1.CosmosFullNode
-	crd.Name = "agoric"
-	crd.Namespace = "test"
+	t.Run("happy path - private addresses", func(t *testing.T) {
+		var crd cosmosv1.CosmosFullNode
+		crd.Name = "dydx"
+		crd.Namespace = namespace
+		crd.Spec.Replicas = 2
+		res, err := BuildNodeKeySecrets(nil, &crd)
+		require.NoError(t, err)
+		secret := res[0].Object()
+		secret.Data[nodeKeyFile] = []byte(nodeKey)
 
-	t.Run("happy path", func(t *testing.T) {
-		lister := mockLister(func(ctx context.Context, list client.ObjectList, opts ...client.ListOption) error {
-			list.(*corev1.PodList).Items = pods
-
-			require.Len(t, opts, 2)
-			var listOpt client.ListOptions
-			for _, opt := range opts {
-				opt.ApplyToList(&listOpt)
+		var (
+			getCount int
+			objKeys  []client.ObjectKey
+		)
+		getter := mockGetter(func(ctx context.Context, key client.ObjectKey, obj client.Object, opts ...client.GetOption) error {
+			objKeys = append(objKeys, key)
+			getCount++
+			switch ref := obj.(type) {
+			case *corev1.Secret:
+				*ref = *secret
+			case *corev1.Service:
+				*ref = corev1.Service{}
 			}
-			require.Equal(t, "test", listOpt.Namespace)
-			require.Zero(t, listOpt.Limit)
-			require.Equal(t, ".metadata.controller=agoric", listOpt.FieldSelector.String())
 			return nil
 		})
 
-		tmClient := mockTmClient(func(ctx context.Context, rpcHost string) (cosmos.TendermintStatus, error) {
-			require.NotNil(t, ctx)
-			var status cosmos.TendermintStatus
-
-			switch rpcHost {
-			case "http://1.1.1.1:26657":
-				status.Result.NodeInfo.ID = "foo"
-				// tcp:// added by tendermint rpc if external_address is blank
-				status.Result.NodeInfo.ListenAddr = "tcp://0.0.0.0:26656"
-				return status, nil
-			case "http://2.2.2.2:26657":
-				status.Result.NodeInfo.ID = "bar"
-				status.Result.NodeInfo.ListenAddr = "12.34.56.78:26656"
-				return status, nil
-			default:
-				panic("unexpected rpcHost: " + rpcHost)
-			}
-		})
-
-		collector := NewPeerCollectorDeprecated(lister, tmClient)
-
-		got, err := collector.CollectAddresses(ctx, &crd)
+		collector := NewPeerCollector(getter)
+		peers, err := collector.Collect(ctx, &crd)
 		require.NoError(t, err)
+		require.Len(t, peers, 2)
 
-		require.Equal(t, []string{"foo@0.0.0.0:26656", "bar@12.34.56.78:26656"}, got)
+		require.Equal(t, 4, getCount) // 2 secrets + 2 services
+
+		wantKeys := []client.ObjectKey{
+			{Name: "dydx-node-key-0", Namespace: namespace},
+			{Name: "dydx-p2p-0", Namespace: namespace},
+			{Name: "dydx-node-key-1", Namespace: namespace},
+			{Name: "dydx-p2p-1", Namespace: namespace},
+		}
+		require.Equal(t, wantKeys, objKeys)
+
+		got := peers[client.ObjectKey{Name: "dydx-0", Namespace: namespace}]
+		require.Equal(t, p2p.ID("1e23ce0b20ae2377925537cc71d1529d723bb892"), got.NodeID)
+		require.Equal(t, "dydx-p2p-0.strangelove.svc.cluster.local:26656", got.PrivateAddress)
+		require.Equal(t, "1e23ce0b20ae2377925537cc71d1529d723bb892@dydx-p2p-0.strangelove.svc.cluster.local:26656", got.PrivatePeer())
+		require.Empty(t, got.ExternalAddress)
+		require.Equal(t, "1e23ce0b20ae2377925537cc71d1529d723bb892@0.0.0.0:26656", got.ExternalPeer())
+
+		got = peers[client.ObjectKey{Name: "dydx-1", Namespace: namespace}]
+		require.NotEmpty(t, got.NodeID)
+		require.Equal(t, "dydx-p2p-1.strangelove.svc.cluster.local:26656", got.PrivateAddress)
+		require.Empty(t, got.ExternalAddress)
+
+		require.False(t, peers.HasIncompleteExternalAddress())
 	})
 
-	t.Run("tendermint error", func(t *testing.T) {
-		lister := mockLister(func(ctx context.Context, list client.ObjectList, opts ...client.ListOption) error {
-			list.(*corev1.PodList).Items = pods
+	t.Run("happy path - external addresses", func(t *testing.T) {
+		var crd cosmosv1.CosmosFullNode
+		crd.Name = "dydx"
+		crd.Namespace = namespace
+		crd.Spec.Replicas = 3
+		res, err := BuildNodeKeySecrets(nil, &crd)
+		require.NoError(t, err)
+		secret := res[0].Object()
+		secret.Data[nodeKeyFile] = []byte(nodeKey)
+
+		getter := mockGetter(func(ctx context.Context, key client.ObjectKey, obj client.Object, opts ...client.GetOption) error {
+			switch ref := obj.(type) {
+			case *corev1.Secret:
+				*ref = *secret
+			case *corev1.Service:
+				var svc corev1.Service
+				switch key.Name {
+				case "dydx-p2p-0":
+					svc.Spec.Type = corev1.ServiceTypeLoadBalancer
+				case "dydx-p2p-1":
+					svc.Spec.Type = corev1.ServiceTypeLoadBalancer
+					svc.Status.LoadBalancer.Ingress = []corev1.LoadBalancerIngress{{IP: "1.2.3.4"}}
+				case "dydx-p2p-2":
+					svc.Spec.Type = corev1.ServiceTypeLoadBalancer
+					svc.Status.LoadBalancer.Ingress = []corev1.LoadBalancerIngress{{Hostname: "host.example.com"}}
+				}
+				*ref = svc
+			}
 			return nil
 		})
 
-		tmClient := mockTmClient(func(ctx context.Context, rpcHost string) (cosmos.TendermintStatus, error) {
-			return cosmos.TendermintStatus{}, errors.New("tendermint error")
-		})
+		collector := NewPeerCollector(getter)
+		peers, err := collector.Collect(ctx, &crd)
+		require.NoError(t, err)
+		require.Len(t, peers, 3)
 
-		collector := NewPeerCollectorDeprecated(lister, tmClient)
+		got := peers[client.ObjectKey{Name: "dydx-0", Namespace: namespace}]
+		require.Equal(t, p2p.ID("1e23ce0b20ae2377925537cc71d1529d723bb892"), got.NodeID)
+		require.Empty(t, got.ExternalAddress)
+		require.Equal(t, "1e23ce0b20ae2377925537cc71d1529d723bb892@0.0.0.0:26656", got.ExternalPeer())
 
-		_, err := collector.CollectAddresses(ctx, &crd)
+		got = peers[client.ObjectKey{Name: "dydx-1", Namespace: namespace}]
+		require.Equal(t, "1.2.3.4:26656", got.ExternalAddress)
+		require.Equal(t, "1e23ce0b20ae2377925537cc71d1529d723bb892@1.2.3.4:26656", got.ExternalPeer())
 
-		require.Error(t, err)
-		require.EqualError(t, err, "tendermint error")
+		got = peers[client.ObjectKey{Name: "dydx-2", Namespace: namespace}]
+		require.Equal(t, "host.example.com:26656", got.ExternalAddress)
+		require.Equal(t, "1e23ce0b20ae2377925537cc71d1529d723bb892@host.example.com:26656", got.ExternalPeer())
+
+		require.True(t, peers.HasIncompleteExternalAddress())
+		want := []string{"1e23ce0b20ae2377925537cc71d1529d723bb892@0.0.0.0:26656",
+			"1e23ce0b20ae2377925537cc71d1529d723bb892@1.2.3.4:26656",
+			"1e23ce0b20ae2377925537cc71d1529d723bb892@host.example.com:26656"}
+		require.ElementsMatch(t, want, peers.AllExternal())
+	})
+
+	t.Run("zero replicas", func(t *testing.T) {
+		var crd cosmosv1.CosmosFullNode
+		crd.Spec.Replicas = 0
+
+		collector := NewPeerCollector(panicGetter)
+		peers, err := collector.Collect(ctx, &crd)
+		require.NoError(t, err)
+		require.Len(t, peers, 0)
 	})
 
 	t.Run("get error", func(t *testing.T) {
-		lister := mockLister(func(ctx context.Context, list client.ObjectList, opts ...client.ListOption) error {
-			return errors.New("list error")
+		getter := mockGetter(func(ctx context.Context, key client.ObjectKey, obj client.Object, opts ...client.GetOption) error {
+			return errors.New("boom")
 		})
 
-		tmClient := mockTmClient(func(ctx context.Context, rpcHost string) (cosmos.TendermintStatus, error) {
-			panic("should not be called")
-		})
-
-		collector := NewPeerCollectorDeprecated(lister, tmClient)
-		_, err := collector.CollectAddresses(ctx, &crd)
+		collector := NewPeerCollector(getter)
+		var crd cosmosv1.CosmosFullNode
+		crd.Name = "dydx"
+		crd.Spec.Replicas = 1
+		_, err := collector.Collect(ctx, &crd)
 
 		require.Error(t, err)
-		require.EqualError(t, err, "list error")
+		require.EqualError(t, err, "get secret dydx-node-key-0: boom")
+		require.True(t, err.IsTransient())
+	})
+
+	t.Run("invalid node key", func(t *testing.T) {
+		getter := mockGetter(func(ctx context.Context, key client.ObjectKey, obj client.Object, opts ...client.GetOption) error {
+			switch ref := obj.(type) {
+			case *corev1.Secret:
+				var secret corev1.Secret
+				secret.Data = map[string][]byte{nodeKeyFile: []byte("invalid")}
+				*ref = secret
+			case *corev1.Service:
+				panic("should not be called")
+			}
+			return nil
+		})
+
+		var crd cosmosv1.CosmosFullNode
+		crd.Name = "dydx"
+		crd.Spec.Replicas = 1
+		collector := NewPeerCollector(getter)
+		_, err := collector.Collect(ctx, &crd)
+
+		require.Error(t, err)
+		require.Contains(t, err.Error(), "invalid character")
+		require.False(t, err.IsTransient())
 	})
 }
